% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_beta_regression_random_effect.R
\name{fit_beta_random_effect}
\alias{fit_beta_random_effect}
\title{Fit beta random effect}
\usage{
fit_beta_random_effect(
  Z = Z,
  Y = Y,
  subject.ind = subject.ind,
  time.ind = time.ind,
  quad.n = 30,
  verbose = FALSE
)
}
\arguments{
\item{Z}{FILL}

\item{Y}{FILL}

\item{subject.ind}{the subject index}

\item{time.ind}{the time index}

\item{quad.n}{number of points in gaussian quadrature}

\item{verbose}{a boolean to enable more output}
}
\value{
a named list
\itemize{
  \item est.table
  \item s2.est
  \item v.est
}
}
\description{
Fit beta random effect
}
