% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_logistic_regression_random_effect.R
\name{fit_logistic_random_effect}
\alias{fit_logistic_random_effect}
\title{Fit logisitic random effect}
\usage{
fit_logistic_random_effect(
  X = X,
  Y = Y,
  subject.ind = subject.ind,
  time.ind = time.ind,
  quad.n = 30,
  verbose = FALSE
)
}
\arguments{
\item{X}{FILL}

\item{Y}{FILL}

\item{subject.ind}{the subject index}

\item{time.ind}{the time index}

\item{quad.n}{number of points in gaussian quadrature}

\item{verbose}{a boolean to enable more output}
}
\value{
a named list
\itemize{
  \item est.table
  \item s1.est
}
}
\description{
Fit logisitic random effect
}
