% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/fit_zero_inflated_beta_regression_random_effect.R
\name{fit_zero_inflated_beta_random_effect}
\alias{fit_zero_inflated_beta_random_effect}
\title{Fit zero inflated beta random effect}
\usage{
fit_zero_inflated_beta_random_effect(
  X = X,
  Z = Z,
  Y = Y,
  subject.ind = subject.ind,
  time.ind = time.ind,
  component.wise.test = TRUE,
  joint.test = TRUE,
  quad.n = 30,
  verbose = FALSE
)
}
\arguments{
\item{X}{FILL}

\item{Z}{FILL}

\item{Y}{FILL}

\item{subject.ind}{the subject index}

\item{time.ind}{the time index}

\item{component.wise.test}{boolean to run component-wise test}

\item{joint.test}{boolean to run joint test}

\item{quad.n}{number of points in gaussian quadrature}

\item{verbose}{a boolean to enable more output}
}
\value{
a named list
\itemize{
  \item logistic.est.table
  \item logistic.s1.est
  \item beta.est.table
  \item beta.s2.est
  \item beta.v.est
  \item loglikelihood
  \item joint.p
}
}
\description{
Fit zero inflated beta random effect
}
