% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_gaussian_quad_points.R
\name{generate_gaussian_quad_points}
\alias{generate_gaussian_quad_points}
\title{Provide a pregenerated gaussian hermitian quad point distribution}
\usage{
generate_gaussian_quad_points(quad_n = 30)
}
\arguments{
\item{quad_n}{the number of points in the distribution (selected from 10, 20, 30, 50 or 100, default 30)}
}
\value{
a structure with a named list
\itemize{
  \item nodes a vector of nodes
  \item weights a vector of weights for nodes
}
}
\description{
Provide a pregenerated gaussian hermitian quad point distribution
}
