% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ibd}
\alias{ibd}
\title{Longitudinal human microbiome data}
\format{
A data frame with 236 rows and 5 variables:
\describe{
  \item{Sample}{Sample IDs}
  \item{Subject}{Subject IDs}
  \item{Time}{Time points}
  \item{Treatment}{Treatment, 0 for antiTNF, 1 for EEN}
  \item{Abundance}{Abundance for Eubacterium}
  ...
}
}
\usage{
ibd
}
\description{
A dataset containing the bacterial abundance and clinical information from
a longitudinal human microbiome study
}
\references{
Lewis and Chen et al. (2016) Cell Host & Microbe 18 (4), 489-500
}
\keyword{datasets}
