% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_beta_data.R
\name{simulate_beta_random_effect_data}
\alias{simulate_beta_random_effect_data}
\title{Simulate beta data}
\usage{
simulate_beta_random_effect_data(
  subject_n = 50,
  time_n = 5,
  v = 2,
  beta = as.matrix(c(-0.5, -0.5, 0.5)),
  Z = NA,
  s2 = 1,
  sim_seed = 100
)
}
\value{
a named list
\itemize{
  \item Y
  \item Z
  \item c
  \item u
  \item v
  \item beta
  \item s2
  \item subject_ind
  \item time_ind
}
}
\description{
Simulate beta data
}
