% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_logistic_data.R
\name{simulate_logistic_data}
\alias{simulate_logistic_data}
\title{Simulate logistic data}
\usage{
simulate_logistic_data()
}
\value{
a named list
\itemize{
  \item X
  \item Y
  \item b
  \item subject_ind
  \item time_ind
}
}
\description{
Simulate logistic data
}
