% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zibr.R
\name{zibr}
\alias{zibr}
\title{Fit zero-inflated beta regression with random effects}
\usage{
zibr(logistic.cov = logistic.cov, beta.cov = beta.cov, Y = Y,
  subject.ind = subject.ind, time.ind = time.ind, quad.n = 30,
  verbose = FALSE)
}
\arguments{
\item{logistic.cov}{the covariates in logistic component}

\item{beta.cov}{the covariates in beta component}

\item{Y}{the response variable in the regression model}

\item{subject.ind}{the variable for subject IDs}

\item{time.ind}{the variable for time points}

\item{quad.n}{Gaussian quadrature points}

\item{verbose}{print the fitting process}
}
\value{
X,Y,Z
}
\description{
Fit zero-inflated beta regression with random effects
}
\examples{
\dontrun{
zibr(logistic.cov=logistic.cov,beta.cov=beta.cov,Y=Y,subject.ind=subject.ind,time.ind=time.ind)
}
}

