% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/fit_zero_inflated_beta_regression_random_effect.R
\name{fit_zero_inflated_beta_random_effect}
\alias{fit_zero_inflated_beta_random_effect}
\title{Fit zero inflated beta random effect}
\usage{
fit_zero_inflated_beta_random_effect(
  X = X,
  Z = Z,
  Y = Y,
  subject_ind = subject_ind,
  time_ind = time_ind,
  component_wise_test = TRUE,
  joint_test = TRUE,
  quad_n = 30,
  verbose = FALSE
)
}
\arguments{
\item{X}{FILL}

\item{Z}{FILL}

\item{Y}{FILL}

\item{subject_ind}{the subject index}

\item{time_ind}{the time index}

\item{component_wise_test}{boolean to run component-wise test}

\item{joint_test}{boolean to run joint test}

\item{quad_n}{number of points in gaussian quadrature}

\item{verbose}{a boolean to enable more output}
}
\value{
a named list
\itemize{
  \item logistic.est.table
  \item logistic.s1.est
  \item beta.est.table
  \item beta.s2.est
  \item beta.v.est
  \item loglikelihood
  \item joint.p
}
}
\description{
Fit zero inflated beta random effect
}
