% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zibr.R
\name{zibr}
\alias{zibr}
\title{Fit zero-inflated beta regression with random effects}
\usage{
zibr(logistic.cov = logistic.cov, beta.cov = beta.cov, Y = Y,
  subject.ind = subject.ind, time.ind = time.ind, quad.n = 30,
  verbose = FALSE)
}
\arguments{
\item{logistic.cov}{the covariates in logistic component}

\item{beta.cov}{the covariates in beta component}

\item{Y}{the response variable in the regression model}

\item{subject.ind}{the variable for subject IDs}

\item{time.ind}{the variable for time points}

\item{quad.n}{Gaussian quadrature points}

\item{verbose}{print the fitting process}
}
\value{
logistic.est.table the estimated coefficients for logistic component.

logistic.s1.est the estimated standard deviation for the random effect in the logistic component.

beta.est.table the estimated coefficients for logistic component.

beta.s2.est the estimated standard deviation for the random effect in the beta component.

beta.v.est the estiamted dispersion parameter in the beta component.

joint.p  the pvalues for jointly testing each covariate in both logistic and beta component.
}
\description{
Fit zero-inflated beta regression with random effects
}
\examples{
\dontrun{
## simulate some data
sim <- simulate_zero_inflated_beta_random_effect_data(
subject.n=100,time.n=5,
X = as.matrix(c(rep(0,50*5),rep(1,50*5))),
Z = as.matrix(c(rep(0,50*5),rep(1,50*5))),
alpha = as.matrix(c(-0.5,1)),
beta = as.matrix(c(-0.5,0.5)),
s1 = 1,s2 = 0.8,
v = 5,
sim.seed=100)
## run zibr on the simulated data
zibr.fit <- zibr(logistic.cov = sim$X, beta.cov = sim$Z, Y = sim$Y,
                 subject.ind = sim$subject.ind,time.ind = sim$time.ind)
zibr.fit
}
}

